fix: source map not update when hmr
