+++ crates/rspack_plugin_hmr/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+edition    = "2021"
+license    = "MIT"
+name       = "rspack_plugin_hmr"
+repository = "https://github.com/web-infra-dev/rspack"
+version    = "0.1.0"
+
+[dependencies]
+rspack_core           = { path = "../rspack_core" }
+rspack_error          = { path = "../rspack_error" }
+rspack_hash           = { path = "../rspack_hash" }
+rspack_identifier     = { path = "../rspack_identifier" }
+rspack_plugin_runtime = { path = "../rspack_plugin_runtime" }
+
+async-trait = { workspace = true }
+rustc-hash  = { workspace = true }
+serde_json  = { workspace = true }
