+++ crates/rspack_plugin_hmr/src/runtime/hot_module_replacement.js
