+++ crates/rspack_plugin_runtime/src/runtime_module/mod.rs
@@ -15,7 +15,6 @@ mod get_trusted_types_policy;
 mod global;
 mod harmony_module_decorator;
 mod has_own_property;
-mod hot_module_replacement;
 mod import_scripts_chunk_loading;
 mod jsonp_chunk_loading;
 mod load_chunk_with_module;
@@ -49,7 +48,6 @@ pub use get_trusted_types_policy::GetTrustedTypesPolicyRuntimeModule;
 pub use global::GlobalRuntimeModule;
 pub use harmony_module_decorator::HarmonyModuleDecoratorRuntimeModule;
 pub use has_own_property::HasOwnPropertyRuntimeModule;
-pub use hot_module_replacement::HotModuleReplacementRuntimeModule;
 pub use import_scripts_chunk_loading::ImportScriptsChunkLoadingRuntimeModule;
 pub use jsonp_chunk_loading::JsonpChunkLoadingRuntimeModule;
 pub use load_chunk_with_module::LoadChunkWithModuleRuntimeModule;
