+++ crates/rspack_testing/Cargo.toml
@@ -28,6 +28,7 @@ rspack_plugin_dev_friendly_split_chunks = { path = "../rspack_plugin_dev_friendl
 rspack_plugin_devtool                   = { path = "../rspack_plugin_devtool" }
 rspack_plugin_entry                     = { path = "../rspack_plugin_entry" }
 rspack_plugin_externals                 = { path = "../rspack_plugin_externals" }
+rspack_plugin_hmr                       = { path = "../rspack_plugin_hmr" }
 rspack_plugin_html                      = { path = "../rspack_plugin_html", features = ["testing"] }
 rspack_plugin_javascript                = { path = "../rspack_plugin_javascript" }
 rspack_plugin_json                      = { path = "../rspack_plugin_json" }
