+++ packages/rspack/tests/hotCases/source-map/disable/webpack.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+  entry: { 
+    main: './index.js',
+  },
+  devtool: false,
+  externalsPresets: {
+    node: true,
+  },
+  node: {
+    __dirname: false,
+  }
+}
