+++ packages/rspack/tests/hotCases/source-map/enable/changed-file.js
@@ -1,0 +1,6 @@
+// TODO: remove this file after cache.
+const path = require('path');
+
+module.exports = [
+  path.resolve(__dirname, './file.js')
+]
