+++ packages/rspack/tests/hotCases/source-map/enable/file.js
@@ -1,0 +1,7 @@
+export default 1;
+---
+export default 2;
+---
+export default 1;
+---
+export default 3;
