+++ packages/rspack/tests/hotCases/source-map/enable/webpack.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+  entry: { 
+    main: './index.js',
+  },
+  devtool: 'cheap-source-map',
+  externalsPresets: {
+    node: true,
+  },
+  node: {
+    __dirname: false,
+  }
+}
